Добавляет исправления по замечаниям
