+++ js/main.js
@@ -1,7 +1,7 @@
 import './form.js';
 import './feed.js';
 import './show-form.js';
-import {createPhotos} from './data.js';
+import {createPhotos} from './create-photos.js';
 // eslint-disable-next-line no-console
 console.log(
   createPhotos()
